// src/spec/tvm-spec.d.ts
export type InstructionName = string
/**
 * Global version (ConfigParam 8) which enables this instruction. Version 9999 means that instruction has no global version and currently unavailable in mainnet.
 */
export type SinceGlobalVersion = number
// ...
export type Aliases = Alias[]

export interface Schema {
    $schema?: string
    /**
     * Instruction list.
     */
    instructions: Instruction[]
    aliases: Aliases
}
export interface Instruction {
    mnemonic: InstructionName
    since_version: SinceGlobalVersion
    doc: Documentation
    bytecode: BytecodeFormat
    value_flow: ValueFlowOfInstruction
    control_flow: ControlFlowOfInstruction
}
/**
 * Free-form human-friendly information which should be used for documentation purposes only.
 */
export interface Documentation {
    opcode?: OpcodeFormatDocumentation
    stack?: StackUsageDescription
    category: CategoryOfInstruction
    description: InstructionDescription
    gas: GasUsageInfo
    fift: FiftUsageDoc
    fift_examples: {
        fift: FiftSnippet
        description: ExampleDescription
    }[]
}
/**
 * Information related to bytecode format of an instruction. Assuming that each instruction has format `prefix || operand_1 || operand_2 || ...` (also some operands may be refs, not bitstring part).
 */
export interface BytecodeFormat {
    tlb: TLBSchema
    prefix: InstructionPrefix
    operands_range_check?: OperandsRangeCheck
    operands: InstructionOperands
}
// ...
export interface ValueFlowOfInstruction {
    inputs: InstructionInputs
    outputs: InstructionOutputs
}
// ...
export interface ControlFlowOfInstruction {
    branches: PossibleBranchesOfAnInstruction
    nobranch: NoBranchPossibility
}
/**
 * Values of saved control flow registers c0-c3
 */
export interface ContinuationSavelist {
    c0?: Continuation
    c1?: Continuation
    c2?: Continuation
    c3?: Continuation
}
export interface Alias {
    mnemonic: AliasName
    alias_of: MnemonicOfAliasedInstruction
    doc_fift?: FiftUsageDoc1
    doc_stack?: StackUsageDescription1
    description?: AliasDescription
    operands: FixedOperandsOfAlias
}